backend/mcp_server.py

# Placeholder for the MCP (Model Context Protocol) Server.
# The 'mcp' library is not standard. This code assumes a hypothetical
# structure based on the user's request. It will need to be adapted
# once the actual 'mcp' library's API is known.

import psycopg2
import os
from dotenv import load_dotenv

# Assume 'mcp' library provides these base classes
# from mcp import MCPServer, BaseTool

# --- Mock/Hypothetical MCP Library Components ---
# If the 'mcp' library is not available, these mock classes
# provide a structure to build upon.

class BaseTool:
    """A mock base class for a tool that can be exposed via MCP."""
    name: str = "Base Tool"
    description: str = "A base tool."

    def run(self, *args, **kwargs):
        raise NotImplementedError("Tool execution not implemented.")

class MCPServer:
    """A mock server to handle tool calls from the LLM."""
    def __init__(self, host="127.0.0.1", port=8090):
        self.host = host
        self.port = port
        self.tools = {}
        print(f"MCP server mock initialized at {self.host}:{self.port}")

    def register_tool(self, tool: BaseTool):
        """Registers a tool to make it available to the LLM."""
        self.tools[tool.name] = tool
        print(f"Tool '{tool.name}' registered with MCP server.")

    def start(self):
        """Starts the server to listen for requests."""
        # In a real implementation, this would start a web server
        # (like FastAPI or Flask) to expose the tool-calling endpoints.
        print("MCP server started (mock).")
        # In a real scenario, this would be a blocking call.
        # e.g., uvicorn.run(self.app, host=self.host, port=self.port)
        pass

# --- Database Connection ---
load_dotenv()
NEON_DATABASE_URL = os.getenv("NEON_DATABASE_URL")

def get_db_connection():
    """Establishes and returns a connection to the Neon Postgres database."""
    if not NEON_DATABASE_URL:
        raise ValueError("NEON_DATABASE_URL environment variable not set.")
    try:
        conn = psycopg2.connect(NEON_DATABASE_URL)
        return conn
    except psycopg2.OperationalError as e:
        print(f"Error connecting to the database: {e}")
        return None

# --- Metadata Tool Definition ---
class BookMetadataTool(BaseTool):
    """A tool to fetch metadata for a book from the Postgres database."""
    name: str = "getBookMetadata"
    description: str = "Fetches metadata for a given book title, such as author, chapters, and publication year."

    def run(self, book_title: str) -> dict:
        """
        The actual logic for the tool.
        Connects to the database and fetches the required information.
        """
        print(f"Executing getBookMetadata tool for title: {book_title}")
        conn = get_db_connection()
        if not conn:
            return {"error": "Database connection failed."}
        
        # This is a sample query. You will need to adapt it to your database schema.
        query = "SELECT author, chapters, publication_year FROM books WHERE title = %s;"
        
        try:
            with conn.cursor() as cur:
                cur.execute(query, (book_title,))
                result = cur.fetchone()
                if result:
                    metadata = {
                        "author": result[0],
                        "chapters": result[1],
                        "publication_year": result[2],
                    }
                    return metadata
                else:
                    return {"error": f"No metadata found for book: {book_title}"}
        except Exception as e:
            return {"error": f"An error occurred while querying the database: {e}"}
        finally:
            conn.close()

# --- Server Initialization ---
def initialize_mcp_server():
    """Creates an instance of the MCP server and registers tools."""
    mcp_server = MCPServer()
    
    # Register all tools that the LLM should have access to
    metadata_tool = BookMetadataTool()
    mcp_server.register_tool(metadata_tool)
    
    return mcp_server

if __name__ == "__main__":
    print("Setting up MCP Server...")
    server = initialize_mcp_server()
    
    # In a real application, you would start the server here.
    # For this boilerplate, we are just demonstrating the setup.
    # For example:
    # server.start()
    
    print("MCP Server setup complete. The server is ready to be started.")
    # Example of running a tool manually
    tool = server.tools.get("getBookMetadata")
    if tool:
        # This will fail if the DB is not set up with the correct table and data.
        # print(tool.run("The Great Gatsby")) 
        pass

